Also pass port 8080 through to Host
It's a bit funny that it worked before this change too. I think it's because of the fact that we expect just a `Twingly::HTTP::ConnectionError` in the specs which happens for basically all types of errors, not just the Timeout class of errors.

I will open an issue concerning improving the specs.

Also document the ports that we pass through.