Add `final_url` to response
This new addition makes it possible to determine the URL that the client ends up on in case of one or more redirects. When using the `follow_redirects` option before, the final URL was found but not accessible from the `Response` object. Faraday's `Env.url` contains the URL of the final request [1].

The specs were updated accordingly.

[1]: https://www.rubydoc.info/gems/faraday/Faraday/Env#url-instance_method

Co-Authored-By: Yen Dinh <70380665+yendi127@users.noreply.github.com>