Fix bug causing the bodies of all requests to be concatenated
When I added the middleware that checks the response body size in #32,
I introduced a bug where the response bodies for all requests in a
redirect "chain" gets included in the body returned in the `Twingly::HTTP::Response`
object.

What I did here was to move the middleware after the FollowRedirects middleware,
to ensure we run the middleware on each request, instead of handling all response
body chunks at once.

The test I added here (along with some other existing tests), would fail with the
following error before this fix:

    1) Twingly::HTTP::Client#post when given a host that redirects when following redirects only returns the body of the final request
      Failure/Error: expect(response.fetch(:body)).to eq("finished!")

        expected: "finished!"
              got: "redirect...finished!"

        (compared using ==)
      Shared Example Group: "common HTTP behaviour for" called from ./spec/lib/twingly/http_spec.rb:554
      # ./spec/lib/twingly/http_spec.rb:159:in `block (5 levels) in <top (required)>'